generalized getFacets
